+++ metametameta/__about__.py
@@ -1,19 +1,19 @@
 """Metadata for metametameta."""
 
 __all__ = [
-    "__title__",
-    "__version__",
-    "__description__",
     "__credits__",
+    "__dependencies__",
+    "__description__",
+    "__keywords__",
     "__readme__",
     "__requires_python__",
-    "__keywords__",
     "__status__",
-    "__dependencies__",
+    "__title__",
+    "__version__",
 ]
 
 __title__ = "metametameta"
-__version__ = "0.1.12"
+__version__ = "0.1.13"
 __description__ = "Generate __about__.py with dunder meta."
 __credits__ = [{"name": "Matthew Martin", "email": "matthewdeanmartin@gmail.com"}]
 __readme__ = "README.md"